backend/requirements.txt
alembic>=1.8.1
apispec>=5.2.2
apscheduler>=3.9.1
atlassian-python-api>=3.25.0
feedparser>=6.0.10
flasgger>=0.9.5
flask>=2.1.3
flask-cors>=3.0.10
flask-jwt-extended>=4.4.2
flask-limiter>=2.5.0
flask-login>=0.6.1
flask-marshmallow>=0.14.0
marshmallow>=3.17.0
marshmallow_sqlalchemy>=0.28.1
pdpyras>=4.5.0
psycopg2>=2.9.3
pyjokes>=0.6.0
python-dotenv>=0.20.0
requests>=2.28.1
slack_bolt>=1.14.1
slack_sdk>=3.18.0
sqlalchemy>=1.4.39
waitress>=2.1.2
werkzeug~=2.0.0
